refactor: extract artist and track routes into dedicated modules
Move 4 artist routes (/artists, /artists/{id}, compare, merge-preview)
to ui/artists.py and 5 track routes (/tracks, /tracks/{id}, compare,
merge-preview, /history) to ui/tracks.py. All routes converted from
inline auth/db patterns to FastAPI dependency injection using the
Session 1 foundation (require_user, get_db, base_context, count_rows,
render_fragment). routes.py reduced from 3711 to 2992 lines.
