+++ src/resonance/app.py
@@ -28,8 +28,10 @@ import resonance.logging as logging_module
 import resonance.middleware.session as session_middleware
 import resonance.models.task as task_models
 import resonance.types as types_module
+import resonance.ui.artists as ui_artists_module
 import resonance.ui.events as ui_events_module
 import resonance.ui.routes as ui_routes_module
+import resonance.ui.tracks as ui_tracks_module
 
 logger = structlog.get_logger()
 
@@ -100,7 +102,9 @@ def create_app() -> fastapi.FastAPI:
     application.include_router(api_v1_module.router)
 
     # Register UI routes
+    application.include_router(ui_artists_module.router)
     application.include_router(ui_events_module.router)
+    application.include_router(ui_tracks_module.router)
     application.include_router(ui_routes_module.router)
 
     # Serve static assets (CSS, JS)
