+++ src/resonance/ui/routes.py
@@ -16,7 +16,6 @@ import sqlalchemy as sa
 import sqlalchemy.ext.asyncio as sa_async
 import sqlalchemy.orm as sa_orm
 
-import resonance.dedup as dedup_module
 import resonance.dependencies as deps_module
 import resonance.merge as merge_module
 import resonance.middleware.session as session_module
@@ -25,7 +24,6 @@ import resonance.models.generator as generator_models
 import resonance.models.music as music_models
 import resonance.models.playlist as playlist_models
 import resonance.models.task as task_models
-import resonance.models.taste as taste_models
 import resonance.models.user as user_models
 import resonance.types as types_module
 import resonance.ui.filters as filters_module
@@ -189,724 +187,7 @@ async def dashboard(
     )
 
 
-@router.get("/artists", response_model=None)
-async def artists_page(
-    request: fastapi.Request,
-    page: int = 1,
-) -> fastapi.responses.HTMLResponse | fastapi.responses.RedirectResponse:
-    """Render paginated artists list with filtering, or redirect to login."""
-    user_id = request.state.session.get("user_id")
-    if not user_id:
-        return fastapi.responses.RedirectResponse(url="/login", status_code=307)
-
-    offset = (page - 1) * _PAGE_SIZE
-
-    # Parse filter parameters
-    params = dict(request.query_params)
-    presets = view_filters_module.ARTIST_PRESETS
-    active_preset = view_filters_module.detect_active_preset(params, presets)
-
-    # Parse filters to get active_filters for template context
-    applied = filters_module.parse_filter_params(
-        view_filters_module.ARTIST_FILTERS, params
-    )
-
-    async with _get_db(request) as db:
-        query = sa.select(music_models.Artist)
-
-        # Apply registered filter fields (name, origin, has_events, has_tracks)
-        query = filters_module.apply_filters(
-            query, view_filters_module.ARTIST_FILTERS, params
-        )
-
-        query = (
-            query.order_by(music_models.Artist.name)
-            .offset(offset)
-            .limit(_PAGE_SIZE + 1)
-        )
-
-        result = await db.execute(query)
-        artists = list(result.scalars().all())
-
-    has_next = len(artists) > _PAGE_SIZE
-    artists = artists[:_PAGE_SIZE]
-
-    # Build filter query string for pagination links
-    filter_qs = view_filters_module.build_filter_query_string(
-        applied.active_filters, view_filters_module.ARTIST_FILTERS
-    )
-    # Include quick search in filter_qs
-    q_value = params.get("q", "").strip()
-    if q_value:
-        if filter_qs:
-            filter_qs += f"&q={q_value}"
-        else:
-            filter_qs = f"q={q_value}"
-
-    # Build flat active_filters dict for the template
-    template_active_filters: dict[str, object] = {}
-    for key, value in applied.active_filters.items():
-        if isinstance(value, dict):
-            for dk, dv in value.items():
-                if dv is not None:
-                    template_active_filters[dk] = str(dv)
-        elif isinstance(value, bool):
-            template_active_filters[key] = value
-        else:
-            template_active_filters[key] = value
-    if q_value:
-        template_active_filters["q"] = q_value
-
-    context = {
-        "user_id": user_id,
-        "user_tz": _user_tz(request),
-        "user_role": _user_role(request),
-        "artists": artists,
-        "page": page,
-        "has_next": has_next,
-        "has_prev": page > 1,
-        "active_filters": template_active_filters,
-        "presets": presets,
-        "filters": view_filters_module.ARTIST_TEMPLATE_FILTERS,
-        "active_preset": active_preset,
-        "list_url": "/artists",
-        "list_target": "#artist-list",
-        "filter_qs": filter_qs,
-    }
-
-    if request.headers.get("HX-Request"):
-        return templates.TemplateResponse(request, "partials/artist_list.html", context)
-    return templates.TemplateResponse(request, "artists.html", context)
-
-
-@router.get("/artists/{artist_id}", response_model=None)
-async def artist_detail_page(
-    request: fastapi.Request,
-    artist_id: uuid.UUID,
-    page: int = 1,
-    section: str = "tracks",
-) -> fastapi.responses.HTMLResponse | fastapi.responses.RedirectResponse:
-    """Render artist detail page with tracks, events, candidates, and duplicates."""
-    user_id = request.state.session.get("user_id")
-    if not user_id:
-        return fastapi.responses.RedirectResponse(url="/login", status_code=307)
-
-    offset = (page - 1) * _PAGE_SIZE
-
-    async with _get_db(request) as db:
-        artist_result = await db.execute(
-            sa.select(music_models.Artist).where(music_models.Artist.id == artist_id)
-        )
-        artist = artist_result.scalar_one_or_none()
-
-        if artist is None:
-            raise fastapi.HTTPException(status_code=404, detail="Artist not found")
-
-        track_count = await _count(
-            db, music_models.Track, music_models.Track.artist_id == artist_id
-        )
-
-        tracks_result = await db.execute(
-            sa.select(music_models.Track)
-            .where(music_models.Track.artist_id == artist_id)
-            .order_by(music_models.Track.title)
-            .offset(offset)
-            .limit(_PAGE_SIZE + 1)
-        )
-        tracks = list(tracks_result.scalars().all())
-        tracks_has_next = len(tracks) > _PAGE_SIZE
-        tracks = tracks[:_PAGE_SIZE]
-        tracks_has_prev = page > 1
-
-        events_result = await db.execute(
-            sa.select(concert_models.EventArtist)
-            .where(concert_models.EventArtist.artist_id == artist_id)
-            .options(
-                sa_orm.joinedload(concert_models.EventArtist.event).joinedload(
-                    concert_models.Event.venue
-                )
-            )
-            .order_by(concert_models.EventArtist.position)
-        )
-        event_artists = list(events_result.scalars().unique().all())
-
-        candidates_result = await db.execute(
-            sa.select(concert_models.EventArtistCandidate)
-            .where(
-                concert_models.EventArtistCandidate.matched_artist_id == artist_id,
-                concert_models.EventArtistCandidate.status
-                == types_module.CandidateStatus.PENDING,
-            )
-            .options(sa_orm.joinedload(concert_models.EventArtistCandidate.event))
-        )
-        candidates = list(candidates_result.scalars().unique().all())
-
-        duplicates_result = await db.execute(
-            sa.select(music_models.Artist).where(
-                sa.func.lower(music_models.Artist.name) == sa.func.lower(artist.name),
-                music_models.Artist.id != artist_id,
-            )
-        )
-        duplicates = list(duplicates_result.scalars().all())
-
-    context = {
-        "user_id": user_id,
-        "user_tz": _user_tz(request),
-        "user_role": _user_role(request),
-        "artist": artist,
-        "tracks": tracks,
-        "track_count": track_count,
-        "tracks_has_next": tracks_has_next,
-        "tracks_has_prev": tracks_has_prev,
-        "event_artists": event_artists,
-        "candidates": candidates,
-        "duplicates": duplicates,
-        "page": page,
-        "section": section,
-    }
-
-    if request.headers.get("HX-Request") and section == "tracks":
-        return templates.TemplateResponse(
-            request, "partials/artist_tracks.html", context
-        )
-    return templates.TemplateResponse(request, "artist_detail.html", context)
-
-
-@router.get("/tracks", response_model=None)
-async def tracks_page(
-    request: fastapi.Request,
-    page: int = 1,
-) -> fastapi.responses.HTMLResponse | fastapi.responses.RedirectResponse:
-    """Render paginated tracks list with artist names and filtering."""
-    user_id = request.state.session.get("user_id")
-    if not user_id:
-        return fastapi.responses.RedirectResponse(url="/login", status_code=307)
-
-    offset = (page - 1) * _PAGE_SIZE
-
-    # Parse filter parameters
-    params = dict(request.query_params)
-    presets = view_filters_module.TRACK_PRESETS
-    active_preset = view_filters_module.detect_active_preset(params, presets)
-
-    # Parse filters to get active_filters for template context
-    applied = filters_module.parse_filter_params(
-        view_filters_module.TRACK_FILTERS, params
-    )
-
-    async with _get_db(request) as db:
-        # Build base query — join Artist for artist name filtering
-        query = sa.select(music_models.Track).join(music_models.Artist)
-
-        # Apply registered filter fields (title, artist, recently_played)
-        query = filters_module.apply_filters(
-            query, view_filters_module.TRACK_FILTERS, params
-        )
-
-        query = (
-            query.order_by(music_models.Track.title)
-            .options(sa_orm.joinedload(music_models.Track.artist))
-            .offset(offset)
-            .limit(_PAGE_SIZE + 1)
-        )
-
-        result = await db.execute(query)
-        tracks = list(result.scalars().unique().all())
-
-    has_next = len(tracks) > _PAGE_SIZE
-    tracks = tracks[:_PAGE_SIZE]
-
-    # Build filter query string for pagination links
-    filter_qs = view_filters_module.build_filter_query_string(
-        applied.active_filters, view_filters_module.TRACK_FILTERS
-    )
-    # Include quick search in filter_qs
-    q_value = params.get("q", "").strip()
-    if q_value:
-        if filter_qs:
-            filter_qs += f"&q={q_value}"
-        else:
-            filter_qs = f"q={q_value}"
-
-    # Build flat active_filters dict for the template
-    template_active_filters: dict[str, object] = {}
-    for key, value in applied.active_filters.items():
-        if isinstance(value, dict):
-            for dk, dv in value.items():
-                if dv is not None:
-                    template_active_filters[dk] = str(dv)
-        elif isinstance(value, bool):
-            template_active_filters[key] = value
-        else:
-            template_active_filters[key] = value
-    if q_value:
-        template_active_filters["q"] = q_value
-
-    context = {
-        "user_id": user_id,
-        "user_tz": _user_tz(request),
-        "user_role": _user_role(request),
-        "tracks": tracks,
-        "page": page,
-        "has_next": has_next,
-        "has_prev": page > 1,
-        "active_filters": template_active_filters,
-        "presets": presets,
-        "filters": view_filters_module.TRACK_TEMPLATE_FILTERS,
-        "active_preset": active_preset,
-        "list_url": "/tracks",
-        "list_target": "#track-list",
-        "filter_qs": filter_qs,
-    }
-
-    if request.headers.get("HX-Request"):
-        return templates.TemplateResponse(request, "partials/track_list.html", context)
-    return templates.TemplateResponse(request, "tracks.html", context)
-
-
-@router.get("/tracks/{track_id}", response_model=None)
-async def track_detail_page(
-    request: fastapi.Request,
-    track_id: uuid.UUID,
-    page: int = 1,
-) -> fastapi.responses.HTMLResponse | fastapi.responses.RedirectResponse:
-    """Render track detail page with listening history and duplicates."""
-    user_id = request.state.session.get("user_id")
-    if not user_id:
-        return fastapi.responses.RedirectResponse(url="/login", status_code=307)
-
-    user_uuid = uuid.UUID(user_id)
-    offset = (page - 1) * _PAGE_SIZE
-
-    async with _get_db(request) as db:
-        track_result = await db.execute(
-            sa.select(music_models.Track)
-            .where(music_models.Track.id == track_id)
-            .options(sa_orm.joinedload(music_models.Track.artist))
-        )
-        track = track_result.scalar_one_or_none()
-
-        if track is None:
-            raise fastapi.HTTPException(status_code=404, detail="Track not found")
-
-        history_result = await db.execute(
-            sa.select(music_models.ListeningEvent)
-            .where(
-                music_models.ListeningEvent.track_id == track_id,
-                music_models.ListeningEvent.user_id == user_uuid,
-            )
-            .order_by(music_models.ListeningEvent.listened_at.desc())
-            .offset(offset)
-            .limit(_PAGE_SIZE + 1)
-        )
-        history = list(history_result.scalars().all())
-        has_next = len(history) > _PAGE_SIZE
-        history = history[:_PAGE_SIZE]
-        has_prev = page > 1
-
-        duplicates_result = await db.execute(
-            sa.select(music_models.Track)
-            .where(
-                sa.func.lower(music_models.Track.title) == sa.func.lower(track.title),
-                music_models.Track.artist_id == track.artist_id,
-                music_models.Track.id != track_id,
-            )
-            .options(sa_orm.joinedload(music_models.Track.artist))
-        )
-        duplicates = list(duplicates_result.scalars().unique().all())
-
-    context = {
-        "user_id": user_id,
-        "user_tz": _user_tz(request),
-        "user_role": _user_role(request),
-        "track": track,
-        "history": history,
-        "duplicates": duplicates,
-        "page": page,
-        "has_next": has_next,
-        "has_prev": has_prev,
-    }
-
-    if request.headers.get("HX-Request"):
-        return templates.TemplateResponse(
-            request, "partials/track_history.html", context
-        )
-    return templates.TemplateResponse(request, "track_detail.html", context)
-
-
-@router.get("/artists/{artist_id}/compare/{other_id}", response_model=None)
-async def artist_compare_page(
-    request: fastapi.Request,
-    artist_id: uuid.UUID,
-    other_id: uuid.UUID,
-) -> fastapi.responses.HTMLResponse | fastapi.responses.RedirectResponse:
-    """Render side-by-side comparison of two artists with merge controls."""
-    user_id = request.state.session.get("user_id")
-    if not user_id:
-        return fastapi.responses.RedirectResponse(url="/login", status_code=307)
-
-    async with _get_db(request) as db:
-        artist_a_result = await db.execute(
-            sa.select(music_models.Artist).where(music_models.Artist.id == artist_id)
-        )
-        artist_a = artist_a_result.scalar_one_or_none()
-
-        artist_b_result = await db.execute(
-            sa.select(music_models.Artist).where(music_models.Artist.id == other_id)
-        )
-        artist_b = artist_b_result.scalar_one_or_none()
-
-        if artist_a is None or artist_b is None:
-            raise fastapi.HTTPException(status_code=404, detail="Artist not found")
-
-        canonical, duplicate = dedup_module.pick_canonical(artist_a, artist_b)
-
-        a_track_count = await _count(
-            db, music_models.Track, music_models.Track.artist_id == artist_id
-        )
-        b_track_count = await _count(
-            db, music_models.Track, music_models.Track.artist_id == other_id
-        )
-        a_event_count = await _count(
-            db,
-            concert_models.EventArtist,
-            concert_models.EventArtist.artist_id == artist_id,
-        )
-        b_event_count = await _count(
-            db,
-            concert_models.EventArtist,
-            concert_models.EventArtist.artist_id == other_id,
-        )
-
-    return templates.TemplateResponse(
-        request,
-        "artist_compare.html",
-        {
-            "user_id": user_id,
-            "user_tz": _user_tz(request),
-            "user_role": _user_role(request),
-            "artist_a": artist_a,
-            "artist_b": artist_b,
-            "canonical": canonical,
-            "duplicate": duplicate,
-            "a_track_count": a_track_count,
-            "b_track_count": b_track_count,
-            "a_event_count": a_event_count,
-            "b_event_count": b_event_count,
-        },
-    )
-
-
-@router.get("/tracks/{track_id}/compare/{other_id}", response_model=None)
-async def track_compare_page(
-    request: fastapi.Request,
-    track_id: uuid.UUID,
-    other_id: uuid.UUID,
-) -> fastapi.responses.HTMLResponse | fastapi.responses.RedirectResponse:
-    """Render side-by-side comparison of two tracks with merge controls."""
-    user_id = request.state.session.get("user_id")
-    if not user_id:
-        return fastapi.responses.RedirectResponse(url="/login", status_code=307)
-
-    user_uuid = uuid.UUID(user_id)
-
-    async with _get_db(request) as db:
-        track_a_result = await db.execute(
-            sa.select(music_models.Track)
-            .where(music_models.Track.id == track_id)
-            .options(sa_orm.joinedload(music_models.Track.artist))
-        )
-        track_a = track_a_result.scalar_one_or_none()
-
-        track_b_result = await db.execute(
-            sa.select(music_models.Track)
-            .where(music_models.Track.id == other_id)
-            .options(sa_orm.joinedload(music_models.Track.artist))
-        )
-        track_b = track_b_result.scalar_one_or_none()
-
-        if track_a is None or track_b is None:
-            raise fastapi.HTTPException(status_code=404, detail="Track not found")
-
-        canonical, duplicate = dedup_module.pick_canonical_track(track_a, track_b)
-
-        a_listen_count = await _count(
-            db,
-            music_models.ListeningEvent,
-            music_models.ListeningEvent.track_id == track_id,
-            music_models.ListeningEvent.user_id == user_uuid,
-        )
-        b_listen_count = await _count(
-            db,
-            music_models.ListeningEvent,
-            music_models.ListeningEvent.track_id == other_id,
-            music_models.ListeningEvent.user_id == user_uuid,
-        )
-
-    return templates.TemplateResponse(
-        request,
-        "track_compare.html",
-        {
-            "user_id": user_id,
-            "user_tz": _user_tz(request),
-            "user_role": _user_role(request),
-            "track_a": track_a,
-            "track_b": track_b,
-            "canonical": canonical,
-            "duplicate": duplicate,
-            "a_listen_count": a_listen_count,
-            "b_listen_count": b_listen_count,
-        },
-    )
-
-
-@router.post("/tracks/{track_id}/merge-preview/{other_id}", response_model=None)
-async def track_merge_preview(
-    request: fastapi.Request,
-    track_id: uuid.UUID,
-    other_id: uuid.UUID,
-) -> fastapi.responses.HTMLResponse | fastapi.responses.RedirectResponse:
-    """Return track merge impact summary partial for HTMX."""
-    user_id = request.state.session.get("user_id")
-    if not user_id:
-        return fastapi.responses.RedirectResponse(url="/login", status_code=307)
-
-    async with _get_db(request) as db:
-        canonical_result = await db.execute(
-            sa.select(music_models.Track).where(music_models.Track.id == track_id)
-        )
-        canonical = canonical_result.scalar_one_or_none()
-
-        duplicate_result = await db.execute(
-            sa.select(music_models.Track).where(music_models.Track.id == other_id)
-        )
-        duplicate = duplicate_result.scalar_one_or_none()
-
-        if canonical is None or duplicate is None:
-            raise fastapi.HTTPException(status_code=404, detail="Track not found")
-
-        events_to_repoint = await _count(
-            db,
-            music_models.ListeningEvent,
-            music_models.ListeningEvent.track_id == other_id,
-        )
-        relations_to_repoint = await _count(
-            db,
-            taste_models.UserTrackRelation,
-            taste_models.UserTrackRelation.track_id == other_id,
-        )
-        playlist_appearances = await _count(
-            db,
-            playlist_models.PlaylistTrack,
-            playlist_models.PlaylistTrack.track_id == other_id,
-        )
-
-        merged_links = dict(canonical.service_links or {})
-        for k, v in (duplicate.service_links or {}).items():
-            if v and k not in merged_links:
-                merged_links[k] = v
-
-        duration_backfill = None
-        if not canonical.duration_ms and duplicate.duration_ms:
-            mins = duplicate.duration_ms // 60000
-            secs = (duplicate.duration_ms % 60000) // 1000
-            duration_backfill = f"{mins}:{secs:02d}"
-
-    return templates.TemplateResponse(
-        request,
-        "partials/track_merge_preview.html",
-        {
-            "canonical": canonical,
-            "duplicate": duplicate,
-            "events_to_repoint": events_to_repoint,
-            "relations_to_repoint": relations_to_repoint,
-            "playlist_appearances": playlist_appearances,
-            "merged_links": merged_links,
-            "duration_backfill": duration_backfill,
-        },
-    )
-
-
-@router.post("/artists/{artist_id}/merge-preview/{other_id}", response_model=None)
-async def artist_merge_preview(
-    request: fastapi.Request,
-    artist_id: uuid.UUID,
-    other_id: uuid.UUID,
-) -> fastapi.responses.HTMLResponse | fastapi.responses.RedirectResponse:
-    """Return merge impact summary partial for HTMX."""
-    user_id = request.state.session.get("user_id")
-    if not user_id:
-        return fastapi.responses.RedirectResponse(url="/login", status_code=307)
-
-    async with _get_db(request) as db:
-        canonical_result = await db.execute(
-            sa.select(music_models.Artist).where(music_models.Artist.id == artist_id)
-        )
-        canonical = canonical_result.scalar_one_or_none()
-
-        duplicate_result = await db.execute(
-            sa.select(music_models.Artist).where(music_models.Artist.id == other_id)
-        )
-        duplicate = duplicate_result.scalar_one_or_none()
-
-        if canonical is None or duplicate is None:
-            raise fastapi.HTTPException(status_code=404, detail="Artist not found")
-
-        tracks_to_repoint = await _count(
-            db, music_models.Track, music_models.Track.artist_id == other_id
-        )
-        events_to_repoint = await _count(
-            db,
-            concert_models.EventArtist,
-            concert_models.EventArtist.artist_id == other_id,
-        )
-
-        merged_links = dict(canonical.service_links or {})
-        for k, v in (duplicate.service_links or {}).items():
-            if v and k not in merged_links:
-                merged_links[k] = v
-
-    return templates.TemplateResponse(
-        request,
-        "partials/merge_preview.html",
-        {
-            "canonical": canonical,
-            "duplicate": duplicate,
-            "tracks_to_repoint": tracks_to_repoint,
-            "events_to_repoint": events_to_repoint,
-            "merged_links": merged_links,
-        },
-    )
-
-
-@router.get("/history", response_model=None)
-async def history_page(
-    request: fastapi.Request,
-    page: int = 1,
-) -> fastapi.responses.HTMLResponse | fastapi.responses.RedirectResponse:
-    """Render paginated listening history with filtering."""
-    user_id = request.state.session.get("user_id")
-    if not user_id:
-        return fastapi.responses.RedirectResponse(url="/login", status_code=307)
-
-    user_uuid = uuid.UUID(user_id)
-    offset = (page - 1) * _PAGE_SIZE
-
-    # Parse filter parameters
-    params = dict(request.query_params)
-    multi_params = {
-        "source": request.query_params.getlist("source"),
-    }
-    presets = view_filters_module.HISTORY_PRESETS
-    active_preset = view_filters_module.detect_active_preset(params, presets)
-
-    # Parse filters to get active_filters for template context
-    applied = filters_module.parse_filter_params(
-        view_filters_module.HISTORY_FILTERS, params
-    )
-
-    async with _get_db(request) as db:
-        # Build base query with joins for track/artist filtering
-        query = (
-            sa.select(music_models.ListeningEvent)
-            .join(
-                music_models.Track,
-                music_models.ListeningEvent.track_id == music_models.Track.id,
-            )
-            .join(
-                music_models.Artist,
-                music_models.Track.artist_id == music_models.Artist.id,
-            )
-            .where(music_models.ListeningEvent.user_id == user_uuid)
-        )
-
-        # Apply registered filter fields (track, artist, date)
-        query = filters_module.apply_filters(
-            query, view_filters_module.HISTORY_FILTERS, params
-        )
-
-        # Handle source filter manually (multiselect outside framework)
-        source_values = multi_params.get("source", [])
-        valid_sources = [
-            v for v in source_values if v in ("SPOTIFY", "LISTENBRAINZ", "LASTFM")
-        ]
-        if valid_sources:
-            query = query.where(
-                music_models.ListeningEvent.source_service.in_(valid_sources)
-            )
-
-        query = (
-            query.order_by(music_models.ListeningEvent.listened_at.desc())
-            .options(
-                sa_orm.joinedload(music_models.ListeningEvent.track).joinedload(
-                    music_models.Track.artist
-                )
-            )
-            .offset(offset)
-            .limit(_PAGE_SIZE + 1)
-        )
-
-        result = await db.execute(query)
-        events = list(result.scalars().unique().all())
-
-    has_next = len(events) > _PAGE_SIZE
-    events = events[:_PAGE_SIZE]
-
-    # Build filter query string for pagination links
-    filter_qs = view_filters_module.build_filter_query_string(
-        applied.active_filters, view_filters_module.HISTORY_FILTERS
-    )
-    # Include source in filter_qs (handled outside the standard fields)
-    if valid_sources:
-        source_parts = [f"source={v}" for v in valid_sources]
-        if filter_qs:
-            filter_qs += "&" + "&".join(source_parts)
-        else:
-            filter_qs = "&".join(source_parts)
-    # Include quick search in filter_qs
-    q_value = params.get("q", "").strip()
-    if q_value:
-        if filter_qs:
-            filter_qs += f"&q={q_value}"
-        else:
-            filter_qs = f"q={q_value}"
-
-    # Build flat active_filters dict for the template
-    template_active_filters: dict[str, object] = {}
-    for key, value in applied.active_filters.items():
-        if isinstance(value, dict):
-            for dk, dv in value.items():
-                if dv is not None:
-                    template_active_filters[dk] = str(dv)
-        else:
-            template_active_filters[key] = value
-    if valid_sources:
-        template_active_filters["source"] = valid_sources
-    if q_value:
-        template_active_filters["q"] = q_value
-
-    context = {
-        "user_id": user_id,
-        "user_tz": _user_tz(request),
-        "user_role": _user_role(request),
-        "events": events,
-        "page": page,
-        "has_next": has_next,
-        "has_prev": page > 1,
-        "active_filters": template_active_filters,
-        "presets": presets,
-        "filters": view_filters_module.HISTORY_TEMPLATE_FILTERS,
-        "active_preset": active_preset,
-        "list_url": "/history",
-        "list_target": "#history-list",
-        "filter_qs": filter_qs,
-    }
-
-    if request.headers.get("HX-Request"):
-        return templates.TemplateResponse(
-            request, "partials/history_list.html", context
-        )
-    return templates.TemplateResponse(request, "history.html", context)
+# Artists, tracks, and history routes moved to ui/artists.py and ui/tracks.py
 
 
 @router.get("/playlists", response_model=None)
